chore(datastore)!: Reorganize + import cleanup
Reorganizes the Datastore package to have a folder structure consistent with Dart best practices and cleans up imports to use strictly absolute imports. This fixes a segmentation fault often seen when running unit tests, e.g. https://app.circleci.com/pipelines/github/aws-amplify/amplify-flutter/7615/workflows/4e156adf-6b36-46be-9f59-08b8a92cf937/jobs/94870
