+++ packages/amplify_datastore/analysis_options.yaml
@@ -1,4 +1,3 @@
 analyzer:
   exclude:
     - example
-    - test
