+++ packages/amplify_datastore/lib/src/amplify_datastore_stream_controller.dart
